Split FrontendAction into domain modules, one handler map each
core/FrontendAction.js was a 2000-line grab bag: the dispatcher plus
thirty handlers from six unrelated domains. The action protocol made the
dispatch table pure data, so it can be assembled from domain modules -
FrontendAction is now the thin registry/dispatcher (~130 lines) merging
one handler map per domain:

  core/actions/ControlCall.js   CONTROL_GLOBAL / CONTROL_BY_ID /
                                BINDING_CALL, the whitelists, and the
                                MESSAGE_TOAST / MESSAGE_BOX display
                                hooks folded in from core/Messages.js
                                (deleted - it only carried those)
  core/actions/Browser.js       history/URL, clipboard, downloads,
                                storage, title/favicon, logout, audio
  core/actions/Launchpad.js     FLP shell: cross-app nav, shell title
  core/actions/Variants.js      smart variant / filter bar wiring
  core/actions/Shortcuts.js     keyboard shortcuts + input mode
  core/actions/ViewOps.js       focus, scrolling, element binding,
                                size limits, OData switch, timers,
                                z2ui5 custom functions
  core/actions/LegacyCustomJs.js  the legacy eF()-string parser and
                                raw-JS fallback, moved out of Server.js

The follow-up runners move with it: Server._runSystemJs/_runCustomJs
become FrontendAction.runSystem/runCustom, so Server.js is HTTP and
roundtrip protocol only, and its response wire-format comment stops
documenting fields the response no longer carries.

responseSuccess now hands the response's own request stamp into
_processAfterRendering, so the display guards compare against the
request the response belongs to instead of whatever is newest once
processing starts (and the until-now unused reqSeq parameter earns
its keep again).

Handler bodies moved verbatim; the merged dispatch table was verified
to cover exactly the pre-split event set (33 events). The test loader
learned an opt-in autoLoad that resolves unstubbed z2ui5/* dependencies
from disk, so the specs exercise the composed dispatch as it ships.

src/01/03 regenerated via npm run app2abap.
