+++ app/webapp/core/FrontendAction.js
@@ -1,1935 +1,48 @@
 sap.ui.define(
   [
-    "sap/m/MessageBox",
-    "sap/m/MessageToast",
-    "sap/ui/core/BusyIndicator",
-    "sap/ui/model/odata/v2/ODataModel",
-    "sap/ui/model/Filter",
-    "sap/ui/model/FilterOperator",
-    "sap/ui/model/Sorter",
-    "sap/m/library",
-    "sap/ui/util/Storage",
-    "z2ui5/core/Router",
+    "z2ui5/core/actions/ControlCall",
+    "z2ui5/core/actions/Browser",
+    "z2ui5/core/actions/Launchpad",
+    "z2ui5/core/actions/Variants",
+    "z2ui5/core/actions/Shortcuts",
+    "z2ui5/core/actions/ViewOps",
+    "z2ui5/core/actions/LegacyCustomJs",
     "z2ui5/core/Lib",
-    "z2ui5/core/Messages",
-    "z2ui5/core/ViewSlots",
     "z2ui5/core/AppState",
   ],
   (
-    MessageBox,
-    MessageToast,
-    BusyIndicator,
-    ODataModel,
-    Filter,
-    FilterOperator,
-    Sorter,
-    mobileLibrary,
-    Storage,
-    Router,
+    ControlCall,
+    Browser,
+    Launchpad,
+    Variants,
+    Shortcuts,
+    ViewOps,
+    LegacyCustomJs,
     Lib,
-    Messages,
-    ViewSlots,
     AppState,
   ) => {
     "use strict";
 
     // ------------------------------------------------------------------
-    // Frontend actions: the handlers behind the controller's eF() entry
-    // point. eF itself stays on View1.controller (its name is part of the
-    // protocol - backend-generated view XML binds events to eB/eF); the
-    // behavior lives here, one function per event. Handlers that need to
-    // reach controller state (destroyPopup, eB, ...) receive the calling
+    // Frontend action dispatch: the handlers behind the controller's eF()
+    // entry point and behind every backend follow-up action. eF itself
+    // stays on View1.controller (its name is part of the protocol -
+    // backend-generated view XML binds events to eB/eF); the behavior
+    // lives in the domain modules under core/actions/, one handler map
+    // per domain, merged here into the one dispatch table. Handlers share
+    // the uniform signature (oController, args); ones that need to reach
+    // controller state (slotAction, eB, ...) receive the calling
     // controller as first argument.
     // ------------------------------------------------------------------
-
-    const _URLHelper = mobileLibrary.URLHelper;
-
-    // Animation duration (ms) mapped to a "smooth" scroll request; 0 means an
-    // instant jump. Shared by every scroll path in evScrollTo.
-    const SMOOTH_SCROLL_MS = 300;
-
-    // SMART_VARIANT_INIT waits for the smart controls to register themselves at
-    // the SmartVariantManagement - that happens once their OData metadata has
-    // loaded, so the wait has to survive a slow service (5s) but must not run
-    // forever when no smart control is there at all.
-    const SMART_VARIANT_INIT_TRIES = 50;
-    const SMART_VARIANT_INIT_DELAY = 100;
-
-    // ------------------------------------------------------------------
-    // Launchpad helpers
-    // ------------------------------------------------------------------
-
-    function withCrossAppNavigator(callback) {
-      const nav = AppState.state.oLaunchpad?.CrossAppNavigator;
-      if (!nav) {
-        Lib.logError("CrossAppNav: not running inside Launchpad");
-        return;
-      }
-      try {
-        callback(nav);
-      } catch (e) {
-        Lib.logError("CrossAppNav: callback failed", e);
-      }
-    }
-
-    // NavContainer navigation (NAV_CONTAINER_TO and the NEST/NEST2/POPUP/
-    // POPOVER variants) is no longer dispatched here: the backend routes those
-    // events through the generic CONTROL_BY_ID path (method "to", the
-    // slot passed as the view), so evControlCallById below handles them.
-
-    // ------------------------------------------------------------------
-    // CONTROL_GLOBAL / CONTROL_BY_ID: call a whitelisted method on a
-    // control (by id) or a global object. The whitelist is the safety
-    // boundary; each entry lists the kind of every positional argument so
-    // string payloads are cast/resolved (never "call anything with
-    // anything"). Scope: imperative methods that have no binding equivalent.
-    // ------------------------------------------------------------------
-
-    // control method -> kinds of its positional args. Args beyond the
-    // declared kinds are dropped; trailing args the caller did not send are
-    // not passed at all (so `open()` stays a true no-arg call).
-    const CONTROL_METHODS = {
-      to: ["controlId", "string"], // target page + optional transitionName
-      back: [],
-      toDetail: ["controlId"], // sap.m.SplitApp/SplitContainer: show a detail page
-      toMaster: ["controlId"], // sap.m.SplitApp/SplitContainer: show a master page
-      backDetail: [], // sap.m.SplitApp/SplitContainer: back in the detail stack
-      backMaster: [], // sap.m.SplitApp/SplitContainer: back in the master stack
-      setMode: ["string"], // sap.m.SplitApp/SplitContainer: SplitAppMode
-      navigateBack: [], // sap.m.QuickView/QuickViewCard: navigate one page back
-      focus: [],
-      scrollToIndex: ["int"],
-      scrollTo: ["int", "int"],
-      open: ["string"], // optional page key (ViewSettingsDialog); PDFViewer/Dialog ignore it
-      close: [],
-      setExpanded: ["bool"],
-      discardProgress: ["controlId"],
-      setNextStep: ["controlId"],
-      goToStep: ["controlId", "bool"], // Wizard: target step + focus flag
-      openBy: ["anchor"], // DatePicker/TimePicker/Menu/MessagePopover... anchored open
-      toggleBy: ["anchor"], // sap.m.Menu/MessagePopover: open anchored if closed, close if open
-      setActivePage: ["controlId"], // sap.m.Carousel
-      expandToLevel: ["int"], // sap.m.Tree / sap.ui.table.TreeTable: expand to N levels
-      collapseAll: [], // sap.m.Tree / sap.ui.table.TreeTable: collapse every node
-      setHiddenInPopin: ["object"], // sap.m.Table: hide columns by importance (JSON array of Priority keys)
-      setSticky: ["object"], // sap.m.ListBase/sap.m.Table: JSON array of sap.m.Sticky keys
-      setSelectedSection: ["controlIdOrNull"], // sap.uxap.ObjectPageLayout: an EMPTY argument clears the association
-      setSelectedItem: ["controlIdOrNull"], // sap.m.List/sap.m.Select/...: an EMPTY argument clears the selection
-      setP13nData: ["object"], // sap.m.p13n.*Panel: JSON array of the panel's items
-
-      css: ["string", "string"], // NOT a UI5 method: set one CSS property on the control's own DOM node
-      enablePostButton: ["bool"], // sap.m.FeedInput: toggle the Post button independent of `enabled`
-      addStyleClass: ["string"], // sap.ui.core.Control: add a CSS style class
-      removeStyleClass: ["string"], // sap.ui.core.Control: remove a CSS style class
-      toggleStyleClass: ["string"], // sap.ui.core.Control: toggle a CSS style class
-      setAsyncURLHandler: ["string"], // sap.m.MessagePopover: name of a URL_POLICY below
-    };
-
-    // sap.m.MessagePopover.setAsyncURLHandler expects a live JS callback that
-    // resolves a promise per message link - a shape no backend payload can
-    // carry, and one round-trip per link would be the wrong ergonomics anyway.
-    // The backend names one of these built-in policies instead, so the
-    // link-gating stays declarative data on the wire (see rule "the frontend is
-    // a thin, data-driven executor").
-    const URL_POLICIES = {
-      ALLOW_ALL: () => true,
-      // the sap.m MessagePopover demo case: in-app links (#/x, /path, ?q=) may
-      // be followed, anything that leaves the app (http:, mailto:, //host) is
-      // disabled in the popover.
-      RELATIVE_ONLY: (url) => !isAbsoluteUrl(url),
-      DENY_ALL: () => false,
-    };
-
-    // CONTROL_METHODS.css writes ONE declaration onto the control's own DOM
-    // node. It exists for the case where a control has no property for the
-    // value at all - sap.m.Page has no `width`, so a sample that resizes its
-    // container (a Slider driving `byId(...).$().width(v + "%")`) has nothing
-    // to bind and no method to call. Where the target DOES have the property,
-    // a bound property stays the correct path (rule "prefer a bindable
-    // property"); this is the residue, not a general styling API.
-    //
-    // The property name is checked against this list so the wire stays a
-    // narrow, declarative contract instead of "write anything anywhere"; the
-    // value goes through CSSOM setProperty, which drops an invalid declaration
-    // (a smuggled second declaration never parses).
-    const CSS_PROPERTIES = [
-      "width",
-      "min-width",
-      "max-width",
-      "height",
-      "min-height",
-      "max-height",
-      "color",
-      "background-color",
-      "font-size",
-      "opacity",
-    ];
-
-    function isAbsoluteUrl(url) {
-      const s = String(url ?? "").trim();
-      // a scheme ("http:", "mailto:", "javascript:") or a protocol-relative
-      // "//host" - everything else resolves against the app's own origin
-      return /^[a-z][a-z0-9+.-]*:/i.test(s) || s.startsWith("//");
-    }
-
-    // A method LISTED above carries explicit arg kinds (and some, like openBy/
-    // toggleBy, get special handling). A method NOT listed is still callable as
-    // long as it is a public control method that is not framework-hostile - so
-    // ordinary setters/toggles (setVisible, enablePostButton, setLayout, ...) work
-    // without enumerating each one here. The denylist protects abap2UI5's own
-    // invariants: nothing that frees or reparents a tracked control, swaps the
-    // model/binding out from under the framework, tampers with event handlers, or
-    // drives the render lifecycle by hand. (The backend is the trusted driver, so
-    // this is a footgun guard, not a security boundary - and control[method] is
-    // still checked to be a function before the call, so a typo just no-ops.)
-    // Named setters/mutators (setVisible, addItem, removeItem, ...) stay
-    // allowed - they are the API the backend legitimately drives. Denied are
-    // the framework-hostile methods: teardown/reparenting (destroy, exit,
-    // setParent, addDependent, placeAt), model/binding swaps (setModel,
-    // setBinding*, bind*/unbind*), event-handler tampering (attach*/detach*,
-    // fireEvent), the render lifecycle (rerender, invalidate) and the GENERIC
-    // reflection mutators (setAggregation/addAggregation/insertAggregation/
-    // removeAggregation/removeAllAggregation and their Association twins),
-    // which take the member NAME as an argument and reparent tracked controls
-    // behind the framework's back.
-    // Two lists, because the two halves need different matching. A generic
-    // reflection mutator has ONE spelling, so it is denied by EXACT name -
-    // matching it by prefix would swallow the named per-aggregation methods
-    // that share those first characters (removeAllSelectedDates,
-    // removeAllItems, destroySpecialDates, ...), which are exactly the API
-    // this allowlist is meant to offer: they detach or destroy children the
-    // control itself owns - the same footprint the already-allowed removeItem
-    // has - and carry no invariant beyond it. The other half is hostile in
-    // every spelling (bindProperty as much as bind), so it stays a PREFIX.
-    // Built from lists (not one long literal) so no single source line is too
-    // long once embedded into the ABAP string constant (trans2abap.js caps
-    // generated lines at 255 chars).
-    const CONTROL_METHOD_DENY_EXACT = [
-      "destroy",
-      "exit",
-      "fireEvent",
-      "clone",
-      "applySettings",
-      "setAggregation",
-      "addAggregation",
-      "insertAggregation",
-      "removeAggregation",
-      "removeAllAggregation",
-      "destroyAggregation",
-      "setAssociation",
-      "addAssociation",
-      "removeAssociation",
-      "removeAllAssociation",
-    ];
-    const CONTROL_METHOD_DENY_PREFIXES = [
-      "_",
-      "bind",
-      "unbind",
-      "attach",
-      "detach",
-      "addDependent",
-      "placeAt",
-      "rerender",
-      "invalidate",
-      "setModel",
-      "setBinding", // covers setBindingContext
-      "setParent",
-    ];
-    const CONTROL_METHOD_DENY = new RegExp(
-      "^(" + CONTROL_METHOD_DENY_PREFIXES.join("|") + ")",
+    const handlers = Object.assign(
+      {},
+      ControlCall.handlers,
+      Browser.handlers,
+      Launchpad.handlers,
+      Variants.handlers,
+      Shortcuts.handlers,
+      ViewOps.handlers,
     );
-    const CONTROL_METHOD_DENY_SET = new Set(CONTROL_METHOD_DENY_EXACT);
-    function isSafeControlMethod(method) {
-      return (
-        typeof method === "string" &&
-        method.length > 0 &&
-        !CONTROL_METHOD_DENY_SET.has(method) &&
-        !CONTROL_METHOD_DENY.test(method)
-      );
-    }
-
-    // global object -> lazy getter + its allowed methods (with arg kinds).
-    const GLOBAL_TARGETS = {
-      // The two message targets carry a `display` hook: the call is routed
-      // through Messages instead of straight at the global, so the option
-      // handling client->message_toast_display( ) / message_box_display( )
-      // relies on - an ONCLOSE event name turned into an eB() round-trip, the
-      // details sanitizing, dependentOn, the toast style class - lives in one
-      // place. A bare app call simply arrives without options.
-      MESSAGE_TOAST: {
-        get: () => MessageToast,
-        methods: { show: ["string"] },
-        display: (oController, method, aArgs, mOptions) =>
-          Messages.showToast(aArgs[0], mOptions, oController),
-      },
-      MESSAGE_BOX: {
-        get: () => MessageBox,
-        // the method IS the box type - the availability check below then
-        // reports a type this UI5 version does not carry
-        methods: {
-          show: ["string"],
-          alert: ["string"],
-          confirm: ["string"],
-          information: ["string"],
-          warning: ["string"],
-          error: ["string"],
-          success: ["string"],
-        },
-        display: (oController, method, aArgs, mOptions) =>
-          Messages.showBox(method, aArgs[0], mOptions, oController),
-      },
-      // Not a UI5 global but the framework's own view-slot registry: the
-      // slots (MAIN, NEST, NEST2, POPUP, POPOVER) are what a backend response
-      // opens, fills and tears down, and every one of those calls is a
-      // SYSTEM follow-up action. `display` and `updateModel` are not
-      // ViewSlots methods - they live on the controller, which owns the
-      // fragment loading and the model - so the hook below routes them there.
-      VIEW_SLOTS: {
-        get: () => ViewSlots,
-        // display takes TWO positional arguments, the slot and the view XML.
-        // Declaring both also keeps it off the template path below, which
-        // would otherwise read the XML as a placeholder value for the slot.
-        methods: {
-          destroy: ["string"],
-          display: ["string", "string"],
-          // takes no slot: which slots are open is the frontend's own
-          // knowledge, so the action only says that the model changed
-          updateModel: [],
-        },
-        display: (oController, method, aArgs, mOptions) =>
-          oController.slotAction(method, aArgs[0], aArgs[1], mOptions),
-        // display and updateModel are not ViewSlots methods at all, so the
-        // "does this UI5 version carry it" check below has nothing to look
-        // at - and nothing to answer either, since none of the three depends
-        // on the UI5 version. The whitelist above is the only gate.
-        virtualMethods: true,
-      },
-      // The browser history / URL. Router computes ONE outcome from the whole
-      // options object - adopt the hash, push a route entry, replace it, or
-      // write the app-state hash - which is why it is one call and not a
-      // field per decision input.
-      ROUTER: {
-        get: () => Router,
-        methods: { sync: [] },
-        display: (oController, method, aArgs, mOptions) =>
-          Router.sync(mOptions, mOptions.id),
-      },
-      BUSY_INDICATOR: {
-        get: () => BusyIndicator,
-        methods: { show: ["int"], hide: [] },
-      },
-      // sap/ui/core/Theming only exists since UI5 1.118, so it must NOT be a
-      // hard dependency (it 404s on 1.71 and kills the whole component load).
-      // Resolve it lazily: on modern UI5 the core has it loaded, on 1.71 the
-      // require returns undefined and the dispatch reports "not available".
-      THEMING: {
-        get: () => sap.ui.require("sap/ui/core/Theming"),
-        methods: { setTheme: ["string"] },
-      },
-      // sap/ui/core/Popup exists on every supported release, but
-      // setWithinArea is @since 1.89 - so resolve the module lazily like
-      // THEMING and let the "not available" guard below report the older
-      // runtime instead of failing the component load.
-      POPUP: {
-        get: () => sap.ui.require("sap/ui/core/Popup"),
-        methods: { setWithinArea: ["within"] },
-      },
-      // sap/ui/core/InvisibleMessage is @since 1.78 and is a SINGLETON: it
-      // renders nothing, so it has no id CONTROL_BY_ID could resolve - a
-      // global target is the only way a backend-driven content change can be
-      // announced to a screen reader. Lazy-require like THEMING so 1.71 hits
-      // the "not available" guard instead of failing the component load.
-      // announce(sText, sMode) with sMode Polite (default) | Assertive.
-      INVISIBLE_MESSAGE: {
-        get: () => {
-          const IM = sap.ui.require("sap/ui/core/InvisibleMessage");
-          return IM ? IM.getInstance() : undefined;
-        },
-        methods: { announce: ["string", "string"] },
-      },
-      // sap/ui/core/Formatting (@since 1.120) carries the global formatting
-      // configuration. Custom currencies are the case an app cannot express
-      // otherwise: the digit count of a currency code is neither a control
-      // property nor something a per-binding formatter can register for the
-      // standard sap.ui.model.type.Currency. The payload is a JSON object -
-      // data the backend owns anyway. Lazy-require like THEMING.
-      FORMATTING: {
-        get: () => sap.ui.require("sap/ui/core/Formatting"),
-        methods: {
-          setCustomCurrencies: ["object"], // { CODE: { digits: n }, ... }
-          addCustomCurrency: ["string", "object"], // code, { digits: n }
-        },
-      },
-    };
-
-    // Cast one raw string argument to the kind the whitelist declared.
-    // `view` (optional) is the slot the owning control was resolved in, so a
-    // controlId argument resolves against the same view first - this keeps
-    // slot-local ids unambiguous (e.g. a NavContainer navigating to one of
-    // its own pages) before falling back to the global lookup.
-    // A control CLONED from an aggregation template has no id the backend can
-    // spell. UI5 mints it as `<templateId>-<parentId>-<index>` - deterministic,
-    // but the parent id carries the VIEW PREFIX the framework assigns at
-    // runtime (`v1--tpl-v1--car-0`), which the backend never sees. So an
-    // aggregation item is addressed positionally instead:
-    //
-    //     "<controlId>/<aggregation>/<index>"     e.g. "carousel/pages/2"
-    //
-    // resolved here, on the client, where both the prefix and the aggregation
-    // are known. This is the equivalent of the UI5 controller idiom
-    // `oCarousel.setActivePage(oCarousel.getPages()[i])`, which no id-based
-    // call can express. A plain id (no slashes) resolves exactly as before.
-    const AGG_ITEM = /^([^/]+)\/([A-Za-z_][\w]*)\/(\d+)$/;
-
-    function resolveControl(raw, view) {
-      const byId = (id) =>
-        (view && ViewSlots.byId(view.toUpperCase(), id)) ||
-        ViewSlots.resolveById(id);
-
-      const m = AGG_ITEM.exec(String(raw ?? ""));
-      if (!m) return byId(raw);
-
-      const owner = byId(m[1]);
-      if (!owner || typeof owner.getAggregation !== "function") {
-        Lib.logError(`aggregation item '${raw}': no control '${m[1]}'`);
-        return null;
-      }
-      const items = owner.getAggregation(m[2]);
-      if (!Array.isArray(items)) {
-        Lib.logError(
-          `aggregation item '${raw}': '${m[2]}' is no multiple aggregation of ${m[1]}`,
-        );
-        return null;
-      }
-      const item = items[Number(m[3])];
-      if (!item) {
-        // out of range is not an error the app can see otherwise - the method
-        // would silently receive undefined and do nothing
-        Lib.logError(
-          `aggregation item '${raw}': ${m[2]} has ${items.length} item(s)`,
-        );
-        return null;
-      }
-      return item;
-    }
-
-    function castArg(kind, raw, view) {
-      switch (kind) {
-        case "int":
-          return Number(raw);
-        case "bool":
-          return raw === "true" || raw === "X" || raw === true;
-        case "controlId":
-          return resolveControl(raw, view);
-        case "controlIdOrNull":
-          // an ASSOCIATION cannot be data-bound, so clearing one
-          // (setSelectedSection(null), setSelectedItem(null)) can only travel
-          // as a method argument - and an EMPTY argument must arrive as null,
-          // not as the `false` castArgAuto would infer. Same "empty means
-          // null" contract as the `within` kind below.
-          if (raw === "" || raw === undefined || raw === null) return null;
-          return resolveControl(raw, view) || null;
-        case "anchor":
-          // anchor argument for openBy-style methods: resolve the control id
-          // and hand over the CONTROL itself, not its DOM element. Every
-          // sap.m openBy accepts a control, and MessagePopover.openBy
-          // dereferences oControl.getParent() on its argument, so a bare DOM
-          // element throws ("getParent is not a function") and the popup never
-          // opens. DatePicker/TimePicker/Menu accept a control just as well,
-          // so a control is the universally-correct anchor.
-          return resolveControl(raw, view);
-        case "within":
-          // sap.ui.core.Popup.setWithinArea: a control id confines every popup
-          // to that control, an EMPTY argument releases the restriction (the
-          // popup is bounded by the window again). Popup.convertWithin()
-          // accepts a sap.ui.core.Element and dereferences its DOM node when a
-          // popup opens, so handing over the CONTROL - not its DOM element -
-          // is what survives a re-render of the area in between.
-          if (raw === "" || raw === undefined || raw === null) return null;
-          return resolveControl(raw, view) || null;
-        case "object":
-          // the backend embeds an argument that starts with { or [ as real
-          // JSON (get_event_client_json), so on that path the value arrives
-          // already parsed; only the legacy eF( ) string form needs parsing.
-          if (raw && typeof raw === "object") return raw;
-          try {
-            return JSON.parse(raw);
-          } catch {
-            return {};
-          }
-        default:
-          return raw;
-      }
-    }
-
-    // Infer the type of an argument for a method with no declared kinds (the
-    // generalized-allowlist path). abap2UI5 sends booleans as the ABAP tokens
-    // 'X'/space, so recognize those; everything else passes through as a string
-    // (UI5 coerces numeric strings for index/number setters). A method that needs
-    // a literal 'X'/'true'/'false' string or a JSON object must be declared in
-    // CONTROL_METHODS with an explicit kind, which overrides this inference.
-    function castArgAuto(raw) {
-      if (raw === "X" || raw === "true") return true;
-      if (raw === "" || raw === " " || raw === "false") return false;
-      return raw;
-    }
-
-    // kinds whose EMPTY value is meaningful (null), so a missing trailing
-    // argument still has to be passed: the backend wire drops a trailing empty
-    // t_arg entry, and "clear this association" is exactly a call whose only
-    // argument is empty.
-    const NULLABLE_KINDS = ["controlIdOrNull"];
-
-    function castArgs(kinds, rawArgs, view) {
-      // kinds === null: unlisted-but-allowed method, infer each arg's type
-      if (kinds === null) return rawArgs.map((raw) => castArgAuto(raw));
-      // only cast args the caller actually sent - padding missing trailing
-      // args would turn open() into open(undefined) and ints into NaN. The one
-      // exception is a nullable kind (see above): pad it so the call carries
-      // an explicit null instead of relying on the control's no-arg handling.
-      let count = rawArgs.length;
-      while (count < kinds.length && NULLABLE_KINDS.includes(kinds[count]))
-        count++;
-      return kinds
-        .slice(0, count)
-        .map((kind, i) => castArg(kind, rawArgs[i], view));
-    }
-
-    // Run fn once the openBy/toggleBy anchor is in the DOM. A control anchor
-    // goes through Lib.whenRendered (immediate if already rendered, otherwise
-    // after its next onAfterRendering); anything else (a bare DOM element, or a
-    // missing anchor) runs fn straight away.
-    function whenAnchorRendered(anchor, oController, fn) {
-      if (anchor && typeof anchor.getDomRef === "function") {
-        Lib.whenRendered(anchor, oController, fn);
-      } else {
-        fn();
-      }
-    }
-
-    // args: [_, id, view, method, ...params]
-    function evControlCallById(oController, args) {
-      const [, id, view, method] = args;
-      let kinds = CONTROL_METHODS[method];
-      if (!kinds) {
-        // not explicitly listed: allow any public, non-hostile control method
-        // (kinds = null -> arg types are inferred), else fail closed.
-        if (!isSafeControlMethod(method)) {
-          Lib.logError(`CONTROL_BY_ID: method '${method}' not allowed`);
-          return;
-        }
-        kinds = null;
-      }
-      const control = view
-        ? ViewSlots.byId(view.toUpperCase(), id)
-        : ViewSlots.resolveById(id);
-      // toggleBy is not a real control method: open the control anchored to
-      // the anchor control if it is closed, close it if it is already open
-      // (mirrors openBy for a press-to-toggle button). The popup's open state lives
-      // client-side, so the decision stays here rather than round-tripping.
-      if (method === "toggleBy") {
-        if (!control || typeof control.openBy !== "function") {
-          Lib.logError(
-            `CONTROL_BY_ID: 'toggleBy' not callable on control '${id}'`,
-          );
-          return;
-        }
-        const anchor = castArgs(kinds, args.slice(4), view)[0];
-        // Defer the open until the anchor is rendered: a Save-style roundtrip
-        // can make the anchor (e.g. a button hidden until there are messages)
-        // visible in the same response, so it may not be in the DOM yet.
-        whenAnchorRendered(anchor, oController, () => {
-          if (control.isOpen?.()) control.close();
-          else control.openBy(anchor);
-        });
-        return;
-      }
-      // css is not a UI5 method either: it writes one whitelisted CSS
-      // declaration onto the control's own DOM node, for the case where the
-      // control has no property carrying that value (sap.m.Page has no
-      // `width`). Like the original jQuery-style samples do, the declaration
-      // lives on the element and is gone after a re-render - the backend
-      // re-sends it with the next view, exactly as it re-sends every property.
-      if (method === "css") {
-        const prop = String(args[4] ?? "").toLowerCase();
-        if (!CSS_PROPERTIES.includes(prop)) {
-          Lib.logError(
-            `CONTROL_BY_ID: css property '${args[4]}' not allowed (allowed: ${CSS_PROPERTIES.join(", ")})`,
-          );
-          return;
-        }
-        const el = control?.getDomRef?.();
-        if (!el) {
-          Lib.logError(`CONTROL_BY_ID: 'css' - control '${id}' has no DOM ref`);
-          return;
-        }
-        el.style.setProperty(prop, String(args[5] ?? ""));
-        return;
-      }
-      // setAsyncURLHandler takes a FUNCTION, so the argument names a policy
-      // (see URL_POLICIES) and the client installs the matching built-in
-      // validator - the wire still carries data, never code.
-      if (method === "setAsyncURLHandler") {
-        const policy = String(args[4] ?? "").toUpperCase();
-        const isAllowed = URL_POLICIES[policy];
-        if (!isAllowed) {
-          Lib.logError(
-            `CONTROL_BY_ID: unknown URL policy '${args[4]}' (allowed: ${Object.keys(URL_POLICIES).join(", ")})`,
-          );
-          return;
-        }
-        if (!control || typeof control.setAsyncURLHandler !== "function") {
-          Lib.logError(
-            `CONTROL_BY_ID: 'setAsyncURLHandler' not callable on control '${id}'`,
-          );
-          return;
-        }
-        control.setAsyncURLHandler((config) => {
-          // the control hands over { url, id, promise }; the promise's
-          // resolve() decides whether the link stays clickable
-          config?.promise?.resolve({
-            allowed: isAllowed(config.url),
-            id: config.id,
-          });
-        });
-        return;
-      }
-      // openBy is handled BEFORE the generic callable check: a control
-      // without an own openBy (sap.ui.unified.Menu) still supports the
-      // anchored open via its open(bWithKeyboard, opener, my, at, of)
-      // signature - the anchor doubles as opener and dock reference.
-      if (method === "openBy") {
-        if (
-          !control ||
-          (typeof control.openBy !== "function" &&
-            typeof control.open !== "function")
-        ) {
-          Lib.logError(
-            `CONTROL_BY_ID: 'openBy' not callable on control '${id}'`,
-          );
-          return;
-        }
-        const anchor = castArgs(kinds, args.slice(4), view)[0];
-        // Same reason as toggleBy: wait for the anchor to render.
-        whenAnchorRendered(anchor, oController, () => {
-          if (typeof control.openBy === "function") control.openBy(anchor);
-          else control.open(false, anchor, "begin top", "begin bottom", anchor);
-        });
-        return;
-      }
-      if (!control || typeof control[method] !== "function") {
-        Lib.logError(
-          `CONTROL_BY_ID: '${method}' not callable on control '${id}'`,
-        );
-        return;
-      }
-      control[method](...castArgs(kinds, args.slice(4), view));
-    }
-
-    // args: [_, object, method, ...params]
-    function evControlCall(oController, args) {
-      const [, name, method] = args;
-      const target = GLOBAL_TARGETS[name];
-      const kinds = target?.methods[method];
-      if (!kinds) {
-        Lib.logError(`CONTROL_GLOBAL: '${name}.${method}' not allowed`);
-        return;
-      }
-      const obj = target.get();
-      if (
-        !obj ||
-        (!target.virtualMethods && typeof obj[method] !== "function")
-      ) {
-        Lib.logError(`CONTROL_GLOBAL: '${name}.${method}' not available`);
-        return;
-      }
-      let raw = args.slice(3);
-      // A `display` target additionally takes an OPTIONS OBJECT as its last
-      // argument (the full set client->message_toast_display( ) sends). It
-      // needs no marker to be told apart from the template values below: the
-      // backend embeds a JSON object argument as real JSON
-      // (get_event_client_json), so it arrives here already parsed, while
-      // every positional and template value is a string.
-      let mOptions;
-      if (target.display) {
-        const last = raw[raw.length - 1];
-        if (last && typeof last === "object") {
-          mOptions = last;
-          raw = raw.slice(0, -1);
-        }
-      }
-      // a single-string method (MessageToast.show, MessageBox.*) may receive
-      // extra positional values: the first arg is then a template and its
-      // {0},{1},... placeholders are replaced by the client-resolved extras, so
-      // a "X has been activated" toast can be composed on the frontend without a
-      // server round-trip. A lone string is passed through unchanged.
-      if (kinds.length === 1 && kinds[0] === "string" && raw.length > 1) {
-        raw = [formatTemplate(String(raw[0]), raw.slice(1))];
-      }
-      if (target.display) {
-        return target.display(oController, method, raw, mOptions || {});
-      }
-      obj[method](...castArgs(kinds, raw));
-    }
-
-    // replace placeholders in a template with the positional values (as
-    // strings). Two forms:
-    //   {N}                -> the Nth value verbatim
-    //   {N?trueText:falseText} -> trueText when the Nth value is truthy, else
-    //                         falseText (a boolean event param arrives as
-    //                         "true"/"false", so a toggle can toast
-    //                         "Pressed"/"Unpressed" without a server round-trip;
-    //                         trueText/falseText carry no ":" or "}")
-    // an out-of-range placeholder is left as-is.
-    function formatTemplate(tpl, values) {
-      return tpl.replace(
-        /\{(\d+)(?:\?([^:}]*):([^}]*))?\}/g,
-        (m, i, tText, fText) => {
-          const n = Number(i);
-          if (n >= values.length) return m;
-          const v = String(values[n]);
-          if (tText === undefined) return v;
-          const truthy = v !== "" && !/^(false|0|undefined|null)$/i.test(v);
-          return truthy ? tText : fText;
-        },
-      );
-    }
-
-    // ------------------------------------------------------------------
-    // BINDING_CALL: apply a declarative filter/sorter to an aggregation
-    // binding of a control resolved by id - the client-side equivalent of
-    // the classic demo kit controller pattern
-    // oList.getBinding("items").filter([new Filter(...)]). Same safety
-    // boundary as CONTROL_BY_ID: only whitelisted binding methods,
-    // only whitelisted filter operators, everything built from data
-    // (path/operator/values), never from code strings.
-    // ------------------------------------------------------------------
-
-    const FILTER_OPERATORS = new Set([
-      "BT",
-      "Contains",
-      "EndsWith",
-      "EQ",
-      "GE",
-      "GT",
-      "LE",
-      "LT",
-      "NB",
-      "NE",
-      "NotContains",
-      "NotEndsWith",
-      "NotStartsWith",
-      "StartsWith",
-    ]);
-
-    const isEmpty = (v) => v == null || v === "";
-
-    // binding method -> builder that turns the trailing params into the
-    // aggregation-update call. A strict whitelist (unlike CONTROL_METHODS,
-    // which now allows any non-denied public control method): an unlisted
-    // binding method fails closed at the lookup.
-    //   filter: params = [path, operator, value1, value2?]
-    //   sort:   params = [path, descending?, group?] (ABAP bools "X"/"")
-    // The backend arg serializer keeps empty args between filled ones as ''
-    // placeholders but trims trailing empties, so all optionals sit at the
-    // end and may arrive as undefined.
-    // Compound form of the filter payload: ONE param carrying a JSON array
-    // of groups, each group an array of [path, operator, value1, value2?]
-    // rows - OR inside a group, AND across groups (the FacetFilter /
-    // ViewSettingsDialog multi-facet shape). Data only: paths, whitelisted
-    // operators and values - never code. An empty groups array clears.
-    function buildFilterGroups(binding, json) {
-      let groups;
-      try {
-        groups = JSON.parse(json);
-      } catch {
-        Lib.logError("BINDING_CALL: malformed filter groups JSON");
-        return;
-      }
-      if (!Array.isArray(groups)) {
-        Lib.logError("BINDING_CALL: filter groups must be an array");
-        return;
-      }
-      groups = groups.filter((g) => Array.isArray(g) && g.length);
-      if (!groups.length) {
-        binding.filter([]);
-        return;
-      }
-      const outer = [];
-      for (const group of groups) {
-        const inner = [];
-        for (const row of group) {
-          const [path, operator, value1, value2] = Array.isArray(row)
-            ? row
-            : [];
-          if (typeof path !== "string" || !FILTER_OPERATORS.has(operator)) {
-            Lib.logError(
-              `BINDING_CALL: bad filter row (path '${path}' / operator '${operator}')`,
-            );
-            return;
-          }
-          inner.push(
-            new Filter(path, FilterOperator[operator], value1, value2),
-          );
-        }
-        outer.push(new Filter(inner, false)); // OR inside the group
-      }
-      binding.filter([new Filter(outer, true)]); // AND across the groups
-    }
-
-    const BINDING_METHODS = {
-      filter(binding, params) {
-        const [path, operator, value1, value2] = params;
-        // A single param that starts with '[' is the compound groups JSON -
-        // a model path can never start with '[', so the sniff is
-        // unambiguous and the positional single-filter form stays as-is.
-        if (
-          params.length === 1 &&
-          typeof path === "string" &&
-          path.trimStart().startsWith("[")
-        ) {
-          buildFilterGroups(binding, path);
-          return;
-        }
-        // No filter values at all -> clear the filter (the demo kit search
-        // pattern: an emptied search field). A one-sided range (empty
-        // value1 but a set value2, e.g. BT with only an upper bound) is a
-        // real filter, so only clear when BOTH values are empty.
-        if (isEmpty(value1) && isEmpty(value2)) {
-          binding.filter([]);
-          return;
-        }
-        if (!FILTER_OPERATORS.has(operator)) {
-          Lib.logError(`BINDING_CALL: operator '${operator}' not allowed`);
-          return;
-        }
-        binding.filter([
-          new Filter(path, FilterOperator[operator], value1, value2),
-        ]);
-      },
-      sort(binding, [path, descending, group]) {
-        binding.sort([
-          new Sorter(path, castArg("bool", descending), castArg("bool", group)),
-        ]);
-      },
-    };
-
-    // args: [_, id, aggregation, method, ...params]
-    function evBindingCall(oController, args) {
-      const [, id, aggregation, method] = args;
-      const build = BINDING_METHODS[method];
-      if (!build) {
-        Lib.logError(`BINDING_CALL: method '${method}' not allowed`);
-        return;
-      }
-      const binding = ViewSlots.resolveById(id)?.getBinding?.(aggregation);
-      if (!binding || typeof binding[method] !== "function") {
-        Lib.logError(
-          `BINDING_CALL: no '${aggregation}' binding with '${method}' on control '${id}'`,
-        );
-        return;
-      }
-      build(binding, args.slice(4));
-    }
-
-    // ------------------------------------------------------------------
-    // Individual event handlers - one per entry in the dispatch table at
-    // the bottom. Uniform signature (oController, args) so the dispatch
-    // stays trivial; handlers that don't need the controller ignore it.
-    // ------------------------------------------------------------------
-
-    function evHistoryBack() {
-      history.back();
-    }
-
-    function evNavToRoute(oController, args) {
-      // Navigate to another app by setting the hash route - the UI5 navTo
-      // equivalent. args[1] is the target app class (or a full "app/<CLASS>"
-      // route). The push adds a browser history entry, so Back returns to the
-      // current app; the router's hashChanged handler then starts the target
-      // app. No-op unless the app enabled routing.
-      const raw = Lib.toText(args[1]);
-      if (!raw) return;
-      Router.navToApp(raw);
-    }
-
-    function evClipboardCopy(oController, args) {
-      Lib.copyToClipboard(args[1]);
-    }
-
-    function evClipboardAppState() {
-      // Guard against a missing response so the copied link never carries
-      // the literal "undefined" as its state id.
-      const id = AppState.state.oResponse?.ID || "";
-      // Router.hrefFor drops the current app hash (e.g. an active app-state)
-      // so the link carries only the fresh state id, but KEEPS the FLP shell
-      // hash - without it the recipient lands on the launchpad home page
-      // instead of this app.
-      Lib.copyToClipboard(Router.hrefFor(`/z2ui5-xapp-state=${id}`));
-    }
-
-    function evDownloadB64File(oController, args) {
-      if (!Lib.isSafeDownloadURL(args[1])) {
-        Lib.logError("DOWNLOAD_B64_FILE: blocked unsafe URL");
-        return;
-      }
-      // A data: URL carrying active HTML combined with an attacker-chosen
-      // .html/.hta filename is a known drive-by vector; block executable data:
-      // MIME types outright (real downloads are octet-stream, images, pdf, ...).
-      if (
-        /^data:(text\/html|application\/xhtml|text\/xml|image\/svg)/i.test(
-          args[1],
-        )
-      ) {
-        Lib.logError("DOWNLOAD_B64_FILE: blocked active data: MIME type");
-        return;
-      }
-      const a = document.createElement("a");
-      a.href = args[1];
-      // Fall back to an empty download attribute when the backend omits the
-      // filename, so the anchor never carries the literal "undefined". Strip
-      // path separators and control characters so the filename cannot escape
-      // the download directory or carry a misleading name.
-      // eslint-disable-next-line no-control-regex -- control chars are matched on purpose here
-      a.download = String(args[2] || "").replace(/[\\/:*?"<>|\x00-\x1f]/g, "_");
-      // Firefox only triggers a programmatic download click when the anchor
-      // is part of the document, so attach it briefly and remove it again.
-      document.body.appendChild(a);
-      a.click();
-      document.body.removeChild(a);
-    }
-
-    function evCrossAppNavToPrevApp() {
-      withCrossAppNavigator((nav) => nav.backToPreviousApp());
-    }
-
-    function evSetSizeLimit(oController, args) {
-      // Two call shapes:
-      //   ["SET_SIZE_LIMIT", "<limit>", "<viewKey>"]   -> set the limit
-      //   ["SET_SIZE_LIMIT", "<viewKey>"]              -> reset the limit
-      const hasLimit = args[2] !== undefined && args[2] !== "";
-      const viewKey = hasLimit ? args[2] : args[1];
-      const limit = hasLimit ? Number(args[1]) : NaN;
-
-      const isValidLimit = Number.isFinite(limit) && limit > 0;
-      if (isValidLimit) {
-        AppState.state.viewSizeLimits[viewKey] = limit;
-      } else {
-        delete AppState.state.viewSizeLimits[viewKey];
-      }
-
-      // MAIN and the two nested views share one root model via propagation, so
-      // resolve the model through MAIN for those slots and apply the effective
-      // (largest) limit across them; popup/popover keep their own model/limit.
-      const modelKey = Lib.isRootModelSlot(viewKey) ? "MAIN" : viewKey;
-      const model = ViewSlots.getView(modelKey)?.getModel();
-      if (model) {
-        const effective = Lib.effectiveSizeLimit(
-          AppState.state.viewSizeLimits,
-          viewKey,
-        );
-        // 100 is the UI5 JSONModel default size limit.
-        model.setSizeLimit(effective ?? 100);
-        model.refresh(true);
-      }
-    }
-
-    function evSetODataModel(oController, args) {
-      let oModel;
-      try {
-        oModel = new ODataModel({
-          serviceUrl: args[1],
-          annotationURI: args[3] || "",
-        });
-        const oView = ViewSlots.getView("MAIN");
-        if (oView) {
-          oView.setModel(oModel, args[2] || undefined);
-        } else {
-          // No view to attach to - release the model instead of leaking it.
-          oModel.destroy();
-        }
-      } catch (e) {
-        Lib.logError(`SET_ODATA_MODEL: failed for '${args[1]}'`, e);
-        // setModel (or the model construction) threw after the model opened
-        // its metadata request - release it so it does not leak.
-        oModel?.destroy?.();
-      }
-    }
-
-    function evStoreData(oController, args) {
-      // Guard against a missing payload so the try below logs a
-      // STORE_DATA-specific error instead of a generic dispatch failure.
-      const { TYPE, PREFIX, VALUE, KEY } = args[1] ?? {};
-      try {
-        const storageType = Storage.Type[TYPE] || Storage.Type.session;
-        const oStorage = new Storage(storageType, PREFIX);
-        if (VALUE === "" || VALUE == null) {
-          oStorage.remove(KEY);
-        } else {
-          oStorage.put(KEY, VALUE);
-        }
-      } catch (e) {
-        Lib.logError(
-          `STORE_DATA: storage operation failed for key '${KEY}'`,
-          e,
-        );
-      }
-    }
-
-    function evCrossAppNavToExt(oController, args) {
-      withCrossAppNavigator((nav) => {
-        const hash =
-          nav.hrefForExternal({ target: args[1], params: args[2] }) || "";
-        if (args[3] === "EXT") {
-          // External redirect: replace the location while keeping the host.
-          // base is the current page (same origin) + a shell-hash fragment,
-          // so this is same-origin by construction; validate anyway to stay
-          // consistent with every other redirect handler in this file.
-          const base = window.location.href.split("#")[0];
-          const url = `${base}${hash}`;
-          if (!Lib.isValidRedirectURL(url)) {
-            Lib.logError(`CrossAppNav EXT: unsafe redirect URL '${url}'`);
-            return;
-          }
-          _URLHelper.redirect(url, true);
-        } else {
-          nav.toExternal({ target: { shellHash: hash } });
-        }
-      });
-    }
-
-    function evLocationReload(oController, args) {
-      if (Lib.isValidRedirectURL(args[1])) {
-        window.location.href = args[1];
-      } else {
-        MessageBox.error(
-          "Invalid redirect URL. Only relative URLs to the same domain are allowed.",
-        );
-      }
-    }
-
-    // SYSTEM_LOGOUT: prefer the launchpad logout when running inside the
-    // FLP; otherwise terminate a possible stateful BSP session first and
-    // then navigate to the logout URL.
-    function evSystemLogout(oController, args) {
-      const logoutUrl = args[1] || "/sap/public/bc/icf/logoff";
-      try {
-        const container = AppState.state.oLaunchpad?.Container;
-        // No explicit logout URL was passed (args is just the event name):
-        // inside the launchpad, prefer its own logout over the BSP/ICF
-        // redirect below.
-        if (container?.logout && args.length <= 1) {
-          container.logout();
-          return;
-        }
-      } catch (e) {
-        Lib.logError("SYSTEM_LOGOUT: ushell logout failed", e);
-      }
-      logoutViaBspTerminate(logoutUrl);
-    }
-
-    // When abap2UI5 is hosted as a BSP application,
-    // /sap/public/bc/icf/logoff alone does not terminate the stateful
-    // BSP context (sap-contextid stays bound to /sap/bc/bsp/sap/<app>/).
-    // Hit the BSP path with ?sap-sessioncmd=logoff first so the BSP
-    // runtime calls server->session->terminate( ), then go to the ICF
-    // logoff to also drop the SSO2 ticket. Outside a BSP path this goes
-    // straight to the logout URL.
-    function logoutViaBspTerminate(logoutUrl) {
-      const path = window.location.pathname;
-      if (!path.startsWith("/sap/bc/bsp/")) {
-        redirectToLogout(logoutUrl);
-        return;
-      }
-
-      // location.pathname never contains a query string, so "?" always starts
-      // the sap-sessioncmd parameter
-      const bspKill = `${path}?sap-sessioncmd=logoff`;
-      let done = false;
-      let frame;
-      const finish = () => {
-        if (done) return;
-        done = true;
-        // Remove the hidden BSP-kill iframe. On a successful logout the page
-        // navigates away and unload cleans up anyway; but if redirectToLogout
-        // blocks an invalid URL (MessageBox, no navigation) the iframe would
-        // otherwise leak - and accumulate over repeated logout attempts.
-        if (frame) {
-          try {
-            frame.remove();
-          } catch {
-            /* already detached */
-          }
-          frame = null;
-        }
-        redirectToLogout(logoutUrl);
-      };
-      try {
-        frame = document.createElement("iframe");
-        frame.style.display = "none";
-        frame.src = bspKill;
-        frame.addEventListener("load", finish);
-        document.body.appendChild(frame);
-      } catch {
-        finish();
-        return;
-      }
-      // Safety net: never wait longer than 1.5s for the BSP terminate.
-      setTimeout(finish, 1500);
-    }
-
-    function redirectToLogout(logoutUrl) {
-      if (Lib.isValidRedirectURL(logoutUrl)) {
-        window.location.href = logoutUrl;
-      } else {
-        MessageBox.error(
-          "Invalid logout URL. Only relative URLs to the same domain are allowed.",
-        );
-      }
-    }
-
-    function evOpenNewTab(oController, args) {
-      if (!Lib.isValidRedirectURL(args[1])) {
-        MessageBox.error(
-          "Invalid URL. Only relative URLs to the same domain are allowed.",
-        );
-        return;
-      }
-      const newWindow = window.open(args[1], "_blank");
-      // Clear opener to prevent the new tab from accessing window.opener.
-      if (newWindow) newWindow.opener = null;
-    }
-
-    // BIND_ELEMENT: element-bind a whole view slot (popup / popover / main) to
-    // a row of a registered table, so the fragment's relative bindings ({Name},
-    // {ProductPicUrl}, ...) resolve against that row - the abap2UI5 equivalent of
-    // oControl.bindElement(oCtx.getPath()). args = [slot, index, path]; the path
-    // comes from client->_bind( table ) (braces already stripped server-side and
-    // again here defensively), the slot from the follow_up_action view param.
-    function evBindElement(oController, args) {
-      const slot = args[1] || "MAIN";
-      const view = ViewSlots.getView(slot);
-      if (!view) {
-        Lib.logError(`BIND_ELEMENT: no view for slot '${slot}'`);
-        return;
-      }
-      const path = String(args[3] ?? "").replace(/[{}]/g, "");
-      if (!path) {
-        Lib.logError("BIND_ELEMENT: empty binding path");
-        return;
-      }
-      view.bindElement(`${path}/${args[2]}`);
-    }
-
-    // The anchor. setPersControler() is sap.ui.comp's own setter and does
-    // more than assign the field: it also creates the control promise that
-    // initialise() insists on. A page variant never gets that call -
-    // addPersonalizableControl() returns early for isPageVariant() and only
-    // the single-control case reaches setPersControler() - which is exactly
-    // why a controller-less app ends up with no anchor and no promise.
-    function anchorPersoControl(oSVM, target) {
-      if (oSVM._oPersoControl) {
-        // a runtime that anchors the control itself is left alone
-      } else if (typeof oSVM.setPersControler === "function") {
-        oSVM.setPersControler(target);
-      } else {
-        // older runtimes without the setter: the field alone still carries
-        // the write path (saving), which is better than nothing
-        oSVM._oPersoControl = target;
-      }
-    }
-
-    // With the anchor in place the load flow still has to be started once.
-    // A smart control does that itself when it registers - but only then,
-    // and it may already have tried (and aborted) before the anchor existed.
-    // So wait for the control's wrapper and initialise it if nobody has:
-    // a wrapper is required (initialise answers "unknown control" without
-    // one) and an initialised wrapper must be left alone ("already executed").
-    function ensureInitialised(oSVM, target, attempt, fnCallback) {
-      if (Lib.isDestroyed(oSVM)) return;
-      const wrapper = oSVM._getControlWrapper
-        ? oSVM._getControlWrapper(target)
-        : null;
-      if (!wrapper) {
-        if (attempt < SMART_VARIANT_INIT_TRIES) {
-          setTimeout(
-            () => ensureInitialised(oSVM, target, attempt + 1, fnCallback),
-            SMART_VARIANT_INIT_DELAY,
-          );
-        }
-        return;
-      }
-      if (wrapper.bInitialized) return;
-      oSVM.initialise(fnCallback || (() => {}), target);
-    }
-
-    // SMART_VARIANT_INIT: place the anchor sap.ui.comp variant management needs
-    // and that only an app controller would otherwise set - the personalizable
-    // control the SmartVariantManagement works against (_oPersoControl).
-    // args = [_, svmId, controlId?].
-    //
-    // Why an anchor and not a call: SmartVariantManagement.initialise(fn, control)
-    // aborts its load flow when `_oPersoControl` is missing ("no personalizable
-    // component available") and marks that control's wrapper as done - a second
-    // call answers "already executed". The smart controls call initialise on
-    // themselves as soon as their metadata arrives, so an anchor set too late
-    // buys nothing: saving works (it only reads the field) but stored variants
-    // are never loaded. Hence: set the field as EARLY as the control exists, and
-    // leave the initialise call to the smart control, which then finds the
-    // anchor in place. Only when no control id was given does this wait for the
-    // registration list and call initialise itself (nobody else will).
-    function evSmartVariantInit(oController, args) {
-      const [, svmId, controlId] = args;
-      let tries = 0;
-      const run = () => {
-        const oSVM = ViewSlots.resolveById(svmId);
-        const control = controlId ? ViewSlots.resolveById(controlId) : null;
-        if (!oSVM || (controlId && !control)) {
-          // the view may still be building - wait for both controls to exist
-          if (tries++ < SMART_VARIANT_INIT_TRIES) {
-            setTimeout(run, SMART_VARIANT_INIT_DELAY);
-            return;
-          }
-          Lib.logError(
-            `SMART_VARIANT_INIT: '${controlId ? `${svmId}' / '${controlId}` : svmId}' not found`,
-          );
-          return;
-        }
-        if (Lib.isDestroyed(oSVM) || typeof oSVM.initialise !== "function") {
-          Lib.logError(
-            `SMART_VARIANT_INIT: no SmartVariantManagement for id '${svmId}'`,
-          );
-          return;
-        }
-        let target = control;
-        if (!target) {
-          // no control named: fall back to the first one that registered, which
-          // means waiting for that registration (it follows the metadata load)
-          const registered = oSVM.getPersonalizableControls
-            ? oSVM.getPersonalizableControls()
-            : [];
-          if (!registered.length) {
-            if (tries++ < SMART_VARIANT_INIT_TRIES) {
-              setTimeout(run, SMART_VARIANT_INIT_DELAY);
-              return;
-            }
-            Lib.logError(
-              `SMART_VARIANT_INIT: no personalizable control registered at '${svmId}'`,
-            );
-            return;
-          }
-          target = ViewSlots.resolveById(registered[0].getControl());
-          if (!target) return;
-        }
-        anchorPersoControl(oSVM, target);
-        ensureInitialised(oSVM, target, 0);
-      };
-
-      run();
-    }
-
-    // ------------------------------------------------------------------
-    // FILTER_BAR_VARIANT_INIT: wire a CLASSIC sap.ui.comp.filterbar.FilterBar
-    // to a SmartVariantManagement. args = [_, svmId, filterBarId].
-    //
-    // A SmartFilterBar registers itself at the variant management (it knows its
-    // fields from the OData metadata), so SMART_VARIANT_INIT above only has to
-    // place the anchor. A classic FilterBar knows nothing about variants: every
-    // list-report controller hand-writes the same three callbacks
-    // (registerFetchData / registerApplyData / registerGetFiltersWithValues),
-    // adds a PersonalizableInfo and marks the variant dirty on each filter
-    // change. That is boilerplate over the bar's own filter items - data, not
-    // app logic - so the framework owns it and the app needs no JavaScript.
-    // ------------------------------------------------------------------
-
-    // marker so a re-sent action cannot stack a second set of callbacks and
-    // change handlers on the same bar (a rebuilt view reuses the ids, but it
-    // builds NEW control instances, which arrive here unmarked)
-    const FILTER_BAR_WIRED = "_z2ui5FilterBarVariantWired";
-
-    function filterItemControl(item) {
-      return item && typeof item.getControl === "function"
-        ? item.getControl()
-        : null;
-    }
-
-    function filterItemValue(item) {
-      const control = filterItemControl(item);
-      return control && typeof control.getValue === "function"
-        ? control.getValue()
-        : "";
-    }
-
-    // what a variant stores, how it is put back, and which filters count as
-    // "assigned" for the bar's own summary text
-    function registerFilterBarCallbacks(oFilterBar) {
-      oFilterBar.registerFetchData(() =>
-        oFilterBar.getAllFilterItems().map((item) => ({
-          groupName: item.getGroupName(),
-          fieldName: item.getName(),
-          fieldData: filterItemValue(item),
-        })),
-      );
-      oFilterBar.registerApplyData((data) => {
-        (data || []).forEach((entry) => {
-          const control = oFilterBar.determineControlByName(
-            entry.fieldName,
-            entry.groupName,
-          );
-          // setValue, not a model write: the two-way binding abap2UI5 put on
-          // the property carries the restored value back to the backend on the
-          // next roundtrip, so selecting a variant needs none of its own
-          if (control && typeof control.setValue === "function") {
-            control.setValue(entry.fieldData);
-          }
-        });
-      });
-      oFilterBar.registerGetFiltersWithValues(() =>
-        oFilterBar
-          .getFilterGroupItems()
-          .filter((item) => String(filterItemValue(item)).length > 0),
-      );
-    }
-
-    // every filter change makes the current variant dirty (the "*" next to the
-    // variant title) and refreshes the bar's assigned-filters summary
-    function attachFilterBarChange(oSVM, oFilterBar) {
-      oFilterBar.getAllFilterItems().forEach((item) => {
-        const control = filterItemControl(item);
-        if (!control || typeof control.attachChange !== "function") return;
-        control.attachChange((oEvent) => {
-          if (typeof oSVM.currentVariantSetModified === "function") {
-            oSVM.currentVariantSetModified(true);
-          }
-          if (typeof oFilterBar.fireFilterChange === "function") {
-            oFilterBar.fireFilterChange(oEvent);
-          }
-        });
-      });
-    }
-
-    // sap.ui.comp is SAPUI5-only, so PersonalizableInfo must never be a hard
-    // dependency of this file (it 404s on OpenUI5 and would kill the component
-    // load). Resolve it at call time: synchronously when the SmartVariant-
-    // Management already pulled it in, asynchronously otherwise.
-    function withPersonalizableInfo(callback) {
-      const name = "sap/ui/comp/smartvariants/PersonalizableInfo";
-      /* ui5lint-disable no-globals --
-       the guard has to read the global sap.ui itself: the point of this
-       function is to load a module that must NOT be a declared dependency
-       (sap.ui.comp is SAPUI5-only and 404s on OpenUI5), so sap.ui.require is
-       the only entry point and there is no injected equivalent to probe. */
-      if (
-        typeof sap === "undefined" ||
-        !sap.ui ||
-        typeof sap.ui.require !== "function"
-      ) {
-        Lib.logError("FILTER_BAR_VARIANT_INIT: sap.ui.require not available");
-        return;
-      }
-      const loaded = sap.ui.require(name);
-      if (loaded) {
-        callback(loaded);
-        return;
-      }
-      sap.ui.require([name], callback, () =>
-        Lib.logError(
-          "FILTER_BAR_VARIANT_INIT: sap.ui.comp.smartvariants not available",
-        ),
-      );
-      /* ui5lint-enable no-globals */
-    }
-
-    function evFilterBarVariantInit(oController, args) {
-      const [, svmId, filterBarId] = args;
-      let tries = 0;
-      const run = () => {
-        const oSVM = ViewSlots.resolveById(svmId);
-        const oFilterBar = ViewSlots.resolveById(filterBarId);
-        if (!oSVM || !oFilterBar) {
-          // the view may still be building - wait for both controls to exist
-          if (tries++ < SMART_VARIANT_INIT_TRIES) {
-            setTimeout(run, SMART_VARIANT_INIT_DELAY);
-            return;
-          }
-          Lib.logError(
-            `FILTER_BAR_VARIANT_INIT: '${svmId}' / '${filterBarId}' not found`,
-          );
-          return;
-        }
-        if (
-          Lib.isDestroyed(oSVM) ||
-          typeof oSVM.addPersonalizableControl !== "function"
-        ) {
-          Lib.logError(
-            `FILTER_BAR_VARIANT_INIT: no SmartVariantManagement for id '${svmId}'`,
-          );
-          return;
-        }
-        if (typeof oFilterBar.registerFetchData !== "function") {
-          Lib.logError(
-            `FILTER_BAR_VARIANT_INIT: no FilterBar for id '${filterBarId}'`,
-          );
-          return;
-        }
-        if (oFilterBar[FILTER_BAR_WIRED]) return;
-        oFilterBar[FILTER_BAR_WIRED] = true;
-
-        registerFilterBarCallbacks(oFilterBar);
-        attachFilterBarChange(oSVM, oFilterBar);
-        withPersonalizableInfo((PersonalizableInfo) => {
-          oSVM.addPersonalizableControl(
-            new PersonalizableInfo({
-              type: "filterBar",
-              keyName: "persistencyKey",
-              dataSource: "",
-              control: oFilterBar,
-            }),
-          );
-          anchorPersoControl(oSVM, oFilterBar);
-          // the load flow ends in registerApplyData, so the bar is clean again
-          // right after it - drop the "*" the restored values would leave
-          ensureInitialised(oSVM, oFilterBar, 0, () => {
-            if (typeof oSVM.currentVariantSetModified === "function") {
-              oSVM.currentVariantSetModified(false);
-            }
-          });
-        });
-      };
-
-      run();
-    }
-
-    function evUrlHelper(oController, args) {
-      const params = args[2] ?? {};
-      // mailto:/sms:/tel: targets are handed to URLHelper as-is; a CR/LF in a
-      // recipient/subject can inject extra headers in some mail clients.
-      // Reject CR/LF in the string params up front.
-      const hasCrLf = (v) => typeof v === "string" && /[\r\n]/.test(v);
-      if (Object.values(params).some(hasCrLf)) {
-        Lib.logError("URLHELPER: blocked CR/LF in parameters");
-        return;
-      }
-      const actions = {
-        REDIRECT: () => {
-          if (!Lib.isSafeRedirectProtocol(params.URL)) {
-            MessageBox.error(
-              "Invalid redirect URL. Only http/https protocols are allowed.",
-            );
-            return;
-          }
-          _URLHelper.redirect(params.URL, params.NEW_WINDOW);
-        },
-        TRIGGER_EMAIL: () =>
-          _URLHelper.triggerEmail(
-            params.EMAIL,
-            params.SUBJECT,
-            params.BODY,
-            params.CC,
-            params.BCC,
-            params.NEW_WINDOW,
-          ),
-        TRIGGER_SMS: () =>
-          _URLHelper.triggerSms(params.TEL, params.TEXT, params.NEW_WINDOW),
-        TRIGGER_TEL: () => _URLHelper.triggerTel(params.TEL),
-      };
-      try {
-        const fn = actions[args[1]];
-        if (fn) fn();
-      } catch (e) {
-        Lib.logError(`URLHELPER: '${args[1]}' failed`, e);
-      }
-    }
-
-    function evImageEditorPopupClose(oController) {
-      let image;
-      try {
-        const editor = ViewSlots.byId("POPUP", "imageEditor");
-        if (editor) image = editor.getImagePngDataURL();
-      } catch (e) {
-        Lib.logError("IMAGE_EDITOR_POPUP_CLOSE: getImagePngDataURL failed", e);
-      }
-      ViewSlots.destroy("POPUP");
-      oController.eB(["SAVE"], image);
-    }
-
-    function evStartTimer(oController, args) {
-      // Intentionally a single timer slot: args[0] is always the event
-      // name "START_TIMER", so a new START_TIMER replaces the previous
-      // one. At most one backend timer is pending at any time - this is
-      // by design, not a bug.
-      const timerKey = args[0];
-      const callbackEvent = args[1];
-      const delay = Number(args[2]) || 0;
-      const timers = AppState.state.timers;
-      clearTimeout(timers[timerKey]);
-      timers[timerKey] = setTimeout(() => {
-        delete timers[timerKey];
-        // dispatch as a background event (args[2] = ignore busy) - a timer
-        // firing while an ordinary roundtrip is in flight must not be
-        // swallowed by the busy guard, or a self-rescheduling poll chain
-        // dies on the first collision with a user click
-        oController.eB([callbackEvent, false, true]);
-      }, delay);
-    }
-
-    // ------------------------------------------------------------------
-    // KEYBOARD_SHORTCUT: bind a key combination to a NAMED BACKEND EVENT -
-    // the declarative equivalent of a sap.ui.core.CommandExecution shortcut
-    // (which needs a controller method and therefore has no place in a
-    // controller-less app). The backend registers "combo -> event" pairs as
-    // data; the document listener below is installed once and always reads
-    // the CURRENT registry, so an app switch (which resets AppState) starts
-    // from an empty set without touching the listener.
-    // ------------------------------------------------------------------
-
-    // in the order they are emitted into a normalized combo, so registration
-    // and keydown produce the same string for any spelling
-    const SHORTCUT_MODIFIERS = ["ctrl", "shift", "alt", "meta"];
-
-    // spellings apps/UI5 use for the same modifier or key
-    const SHORTCUT_ALIASES = {
-      control: "ctrl",
-      cmd: "meta",
-      command: "meta",
-      option: "alt",
-      esc: "escape",
-      del: "delete",
-      ins: "insert",
-      return: "enter",
-      space: " ",
-    };
-
-    function shortcutToken(part) {
-      const t = part.trim().toLowerCase();
-      return SHORTCUT_ALIASES[t] ?? t;
-    }
-
-    // "Ctrl+Shift+S" / "shift + CTRL + s" -> "ctrl+shift+s". Returns an empty
-    // string when no actual key (only modifiers) is named.
-    function normalizeShortcut(combo) {
-      const parts = String(combo ?? "")
-        .split("+")
-        .map(shortcutToken)
-        .filter((p) => p !== "");
-      const mods = SHORTCUT_MODIFIERS.filter((m) => parts.includes(m));
-      const keys = parts.filter((p) => !SHORTCUT_MODIFIERS.includes(p));
-      if (keys.length === 0) return "";
-      return [...mods, keys[keys.length - 1]].join("+");
-    }
-
-    // the same normalized form for an actual keydown event
-    function shortcutFromEvent(oEvent) {
-      const key = String(oEvent.key ?? "").toLowerCase();
-      // a bare modifier press is not a shortcut
-      if (key === "" || SHORTCUT_MODIFIERS.includes(shortcutToken(key)))
-        return "";
-      const mods = [];
-      if (oEvent.ctrlKey) mods.push("ctrl");
-      if (oEvent.shiftKey) mods.push("shift");
-      if (oEvent.altKey) mods.push("alt");
-      if (oEvent.metaKey) mods.push("meta");
-      return [...mods, key].join("+");
-    }
-
-    // A shortcut may be SCOPED, which is how UI5's own CommandExecution
-    // behaves: one in a Popover's dependents shadows the page-level one for
-    // the same command while that popover is open. A scope is either
-    //
-    //   a VIEW SLOT   - POPOVER/POPUP/NEST2/NEST/MAIN, open when the framework
-    //                   has that slot showing (popover_display, popup_display,
-    //                   a nested view)
-    //   a CONTROL ID  - any control DECLARED IN THE VIEW that can be open or
-    //                   closed: a sap.m.Popover/Dialog in `dependents` opened
-    //                   with control_by_id openBy, which is the shape the demo
-    //                   kit's Commands sample actually uses. It never enters a
-    //                   framework slot, so the slot form alone would never fire.
-    //
-    // Dispatch prefers a CONTROL scope (the more specific statement) over a
-    // slot scope, then takes the innermost open slot, then the unscoped entry.
-    const SHORTCUT_SLOTS = ["POPOVER", "POPUP", "NEST2", "NEST", "MAIN"];
-    const SHORTCUT_GLOBAL = ""; // the unscoped registration
-
-    // A control scope counts while the control is OPEN - isOpen() for the
-    // popup-like controls this is for, visibility otherwise.
-    function scopeControlOpen(id) {
-      const c = ViewSlots.resolveById(id);
-      if (!c) return false;
-      if (typeof c.isOpen === "function") return !!c.isOpen();
-      return typeof c.getVisible === "function"
-        ? c.getVisible() !== false
-        : true;
-    }
-
-    function shortcutEntry(combo) {
-      const scopes = AppState.state.shortcuts[combo];
-      if (!scopes) return undefined;
-      for (const key of Object.keys(scopes)) {
-        if (key === SHORTCUT_GLOBAL || SHORTCUT_SLOTS.includes(key)) continue;
-        if (scopeControlOpen(key)) return scopes[key];
-      }
-      for (const key of SHORTCUT_SLOTS) {
-        if (scopes[key] && ViewSlots.getView(key)) return scopes[key];
-      }
-      return scopes[SHORTCUT_GLOBAL];
-    }
-
-    let shortcutListener = null;
-
-    function installShortcutListener() {
-      if (shortcutListener || typeof document === "undefined") return;
-      shortcutListener = (oEvent) => {
-        try {
-          const entry = shortcutEntry(shortcutFromEvent(oEvent));
-          if (!entry) return;
-          // the browser's own default for the combo (Ctrl+S saves the page,
-          // Ctrl+D bookmarks it) must not fire alongside the app command
-          oEvent.preventDefault();
-          entry.controller.eB([entry.event]);
-        } catch (e) {
-          Lib.logError("KEYBOARD_SHORTCUT: dispatch failed", e);
-        }
-      };
-      document.addEventListener("keydown", shortcutListener);
-    }
-
-    // args: [_, combo, eventName, scope] - an empty event name unregisters the
-    // combo IN THAT SCOPE; scope is a view slot key (cs_view-popover/popup/...)
-    // and defaults to the unscoped, always-eligible registration
-    function evKeyboardShortcut(oController, args) {
-      const combo = normalizeShortcut(args[1]);
-      if (!combo) {
-        Lib.logError(
-          `KEYBOARD_SHORTCUT: '${args[1]}' names no key to bind (modifiers only?)`,
-        );
-        return;
-      }
-      // a slot key is matched case-insensitively; anything else is taken as a
-      // control id and keeps its case, because that is how it must resolve
-      const raw = String(args[3] ?? "");
-      const scope = SHORTCUT_SLOTS.includes(raw.toUpperCase())
-        ? raw.toUpperCase()
-        : raw;
-      const shortcuts = AppState.state.shortcuts;
-      const scopes = shortcuts[combo] ?? (shortcuts[combo] = {});
-      if (!args[2]) {
-        delete scopes[scope];
-        // a combo with no registration left must not keep an empty entry:
-        // shortcutEntry would still find it and fall through to undefined,
-        // but preventDefault has already been decided by then
-        if (Object.keys(scopes).length === 0) delete shortcuts[combo];
-        return;
-      }
-      // re-registering a combo in the same scope replaces it, so the backend
-      // can rebind a shortcut without unregistering it first
-      scopes[scope] = { event: args[2], controller: oController };
-      installShortcutListener();
-    }
-
-    function evSetInputMode(oController, args) {
-      try {
-        const oElement = ViewSlots.byId("MAIN", args[1]);
-        if (!oElement) return;
-        const dom = oElement.getDomRef();
-        if (!dom) return;
-        const input = dom.matches("input, textarea")
-          ? dom
-          : dom.querySelector("input, textarea");
-        if (!input) return;
-        input.setAttribute("inputmode", args[2] || "text");
-      } catch (e) {
-        Lib.logError(
-          `KEYBOARD_SET_MODE: setAttribute failed for '${args[1]}'`,
-          e,
-        );
-      }
-    }
-
-    // The three handlers below resolve their target with ViewSlots.resolveById
-    // (not byId "MAIN"): it searches every open slot first, so controls in a
-    // popup/popover/nested view are found, and falls back to the global
-    // registry, so a fully-qualified id resolves too - ids that come from a
-    // UI5 Message (getControlIds()) or any event carry the view prefix.
-
-    function evSetFocus(oController, args) {
-      const oElement = ViewSlots.resolveById(args[1]);
-      if (!oElement) return;
-
-      const applyFocus = () => {
-        try {
-          const info = oElement.getFocusInfo();
-          if (args[2] != null && args[2] !== "") {
-            info.selectionStart = Number(args[2]);
-          }
-          if (args[3] != null && args[3] !== "") {
-            info.selectionEnd = Number(args[3]);
-          }
-          oElement.applyFocusInfo(info);
-        } catch (e) {
-          Lib.logError(`SET_FOCUS: failed for '${args[1]}'`, e);
-        }
-      };
-
-      // The control may still be missing from the DOM when SET_FOCUS runs
-      // together with a fresh view build. Apply now if it is rendered,
-      // otherwise once it is.
-      Lib.whenRendered(oElement, oController, () => {
-        applyFocus();
-        const dom = oElement.getDomRef();
-        if (dom && dom.contains(document.activeElement)) return;
-        // The focus did not stick. A view_model_update in the same response
-        // may have changed the control - e.g. re-enabled a locked input via
-        // its `enabled` binding: the control already reports the new state,
-        // but the DOM still carries the OLD rendering until UI5's async
-        // re-render, and the browser silently ignores focus() on a disabled
-        // element. Re-apply once after the pending re-render has replaced
-        // the DOM.
-        const prevActive = document.activeElement;
-        // "Same place" by node OR by element id: when the re-render also
-        // rebuilt the element that held the focus (the pressed button in the
-        // same form), the focus sits on a NEW node of the SAME control
-        // afterwards - that still counts as "the user did not move it".
-        const samePlace = (el) =>
-          el == null ||
-          el === document.body ||
-          el === prevActive ||
-          Boolean(el.id && prevActive && el.id === prevActive.id);
-        const delegate = {
-          onAfterRendering: () => {
-            oElement.removeEventDelegate(delegate);
-            // Defer past the rendering task: when the re-render replaced the
-            // focused element, UI5's FocusHandler restores its focus AFTER
-            // all onAfterRendering delegates ran - focusing here would be
-            // overridden right away.
-            setTimeout(() => {
-              if (Lib.isDestroyed(oController)) return;
-              // Only when the focus was not actively moved elsewhere in
-              // between - a re-render at some arbitrary later point must
-              // never steal the user's focus.
-              if (!samePlace(document.activeElement)) return;
-              applyFocus();
-            }, 0);
-          },
-        };
-        oElement.addEventDelegate(delegate);
-      });
-    }
-
-    function evScrollTo(oController, args) {
-      // args[1] = control id
-      // args[2] = scrollTop  (Y, vertical, px)
-      // args[3] = scrollLeft (X, horizontal, px) - optional, default 0
-      // args[4] = behavior - "auto" (default) | "smooth" | "instant"
-      // Strategy: prefer the control's scroll delegate (sap.m.Page,
-      // ScrollContainer etc. expose ScrollEnablement). The delegate knows
-      // the real scroll container, which often is NOT the control's root
-      // DOM element - so native Element.scrollTo on getDomRef() silently
-      // does nothing on a Page. ScrollEnablement.scrollTo(x, y, time)
-      // animates when time > 0, so "smooth" maps to a 300ms animation.
-      // Native Element.scrollTo is only used as a fallback for controls
-      // without a delegate.
-      try {
-        const oElement = ViewSlots.resolveById(args[1]);
-        if (!oElement) return;
-        const y = Number(args[2]) || 0;
-        const x = Number(args[3]) || 0;
-        const behavior = args[4] || "auto";
-        const smooth = behavior === "smooth";
-
-        let handled = false;
-        try {
-          const delegate = oElement.getScrollDelegate?.();
-          if (delegate?.scrollTo) {
-            // ScrollEnablement / iScroll delegate: scrollTo(x, y, time)
-            delegate.scrollTo(x, y, smooth ? SMOOTH_SCROLL_MS : 0);
-            handled = true;
-          }
-        } catch {
-          // fall through
-        }
-
-        if (!handled) {
-          const dom =
-            document.getElementById(`${oElement.getId()}-inner`) ||
-            oElement.getDomRef();
-          if (dom?.scrollTo) {
-            dom.scrollTo({ top: y, left: x, behavior });
-          } else if (dom) {
-            dom.scrollTop = y;
-            dom.scrollLeft = x;
-          } else if (oElement.scrollTo) {
-            // sap.m.Page.scrollTo(y, time) - vertical only
-            oElement.scrollTo(y, smooth ? SMOOTH_SCROLL_MS : 0);
-          }
-        }
-      } catch (e) {
-        Lib.logError(`SCROLL_TO: failed for '${args[1]}'`, e);
-      }
-    }
-
-    function evScrollIntoView(oController, args) {
-      // args[1] = control id
-      // args[2] = behavior - "smooth" (default) | "auto" | "instant"
-      // args[3] = block    - "start"  (default) | "center" | "end" | "nearest"
-      // args[4] = inline   - "nearest" (default)| "start"  | "center" | "end"
-      // Modern declarative scroll: bring a control into the viewport,
-      // regardless of where the surrounding scroll container currently is.
-      try {
-        const oElement = ViewSlots.resolveById(args[1]);
-        if (!oElement) return;
-        const dom = oElement.getDomRef();
-        if (!dom || !dom.scrollIntoView) return;
-        dom.scrollIntoView({
-          behavior: args[2] || "smooth",
-          block: args[3] || "start",
-          inline: args[4] || "nearest",
-        });
-      } catch (e) {
-        Lib.logError(`SCROLL_INTO_VIEW: failed for '${args[1]}'`, e);
-      }
-    }
-
-    function evSetTitle(oController, args) {
-      const title = Lib.toText(args[1]);
-      try {
-        document.title = title;
-      } catch (e) {
-        Lib.logError("SET_TITLE: setting document.title failed", e);
-      }
-    }
-
-    function evSetFavicon(oController, args) {
-      const href = Lib.toText(args[1]);
-      try {
-        // Reuse the icon link the page already has instead of appending a
-        // second one - which of two competing <link rel="icon"> elements the
-        // browser honours is up to it, and an app that switches its icon
-        // would otherwise leave one behind per change. `~=` matches one entry
-        // of the whitespace-separated rel list, so a page declaring the
-        // legacy rel="shortcut icon" is found too.
-        const existing = document.head.querySelector('link[rel~="icon"]');
-        if (existing) {
-          existing.href = href;
-          return;
-        }
-        const link = document.createElement("link");
-        link.rel = "icon";
-        link.href = href;
-        document.head.appendChild(link);
-      } catch (e) {
-        Lib.logError("SET_FAVICON: setting the favicon failed", e);
-      }
-    }
-
-    function evSetTitleLaunchpad(oController, args) {
-      const title = Lib.toText(args[1]);
-      try {
-        const shell = AppState.state.oLaunchpad?.ShellUIService;
-        if (shell?.setTitle) {
-          const result = shell.setTitle(title);
-          if (result?.catch) {
-            result.catch((e) =>
-              Lib.logError(
-                "SET_TITLE_LAUNCHPAD: ShellUIService.setTitle failed",
-                e,
-              ),
-            );
-          }
-        }
-      } catch (e) {
-        Lib.logError("SET_TITLE_LAUNCHPAD: ShellUIService.setTitle failed", e);
-      }
-    }
-
-    function evZ2ui5Custom(oController, args) {
-      try {
-        // Custom functions are registered by apps on the public z2ui5
-        // global (js_loader popup), so resolve them via the facade.
-        const fn = AppState.getGlobal(args[1]);
-        if (typeof fn === "function") {
-          fn(args.slice(2));
-        } else {
-          // Missing or not callable (e.g. the app never registered it via
-          // the js_loader popup) - log it instead of failing silently or
-          // with a generic TypeError.
-          Lib.logError(`Z2UI5: 'z2ui5.${args[1]}' is not a function`);
-        }
-      } catch (e) {
-        Lib.logError(`Z2UI5: '${args[1]}' failed`, e);
-      }
-    }
-
-    function evWizardSetNextStep(oController, args) {
-      try {
-        const wiz = ViewSlots.byId("MAIN", args[1]);
-        const step = ViewSlots.byId("MAIN", args[2]);
-        const nextStep = ViewSlots.byId("MAIN", args[3]);
-        if (wiz && step) wiz.discardProgress(step);
-        if (step && nextStep) step.setNextStep(nextStep);
-      } catch (e) {
-        Lib.logError(`WIZARD_SET_NEXT_STEP: failed for wizard '${args[1]}'`, e);
-      }
-    }
-
-    function evPlayAudio(oController, args) {
-      // Only http(s)/data:/blob: sources are meaningful for Audio; validating
-      // the protocol keeps this consistent with the other URL-consuming
-      // actions and blocks odd schemes early.
-      if (!Lib.isSafeDownloadURL(args[1])) {
-        Lib.logError("PLAY_AUDIO: blocked unsafe audio URL");
-        return;
-      }
-      try {
-        const playing = new Audio(args[1]).play();
-        // play() returns a Promise; a rejection (e.g. blocked by the
-        // browser's autoplay policy) is not caught by the surrounding
-        // try/catch and would surface as an unhandled rejection.
-        if (playing?.catch) {
-          playing.catch((e) =>
-            Lib.logError(`PLAY_AUDIO: failed for '${args[1]}'`, e),
-          );
-        }
-      } catch (e) {
-        Lib.logError(`PLAY_AUDIO: failed for '${args[1]}'`, e);
-      }
-    }
-
-    // Frontend event dispatch: maps the eF event name to its handler.
-    const handlers = {
-      SET_SIZE_LIMIT: evSetSizeLimit,
-      HISTORY_BACK: evHistoryBack,
-      NAV_TO_ROUTE: evNavToRoute,
-      CLIPBOARD_COPY: evClipboardCopy,
-      CLIPBOARD_APP_STATE: evClipboardAppState,
-      SET_ODATA_MODEL: evSetODataModel,
-      STORE_DATA: evStoreData,
-      DOWNLOAD_B64_FILE: evDownloadB64File,
-      CROSS_APP_NAV_TO_PREV_APP: evCrossAppNavToPrevApp,
-      CROSS_APP_NAV_TO_EXT: evCrossAppNavToExt,
-      LOCATION_RELOAD: evLocationReload,
-      SYSTEM_LOGOUT: evSystemLogout,
-      OPEN_NEW_TAB: evOpenNewTab,
-      BIND_ELEMENT: evBindElement,
-      URLHELPER: evUrlHelper,
-      IMAGE_EDITOR_POPUP_CLOSE: evImageEditorPopupClose,
-      SET_TITLE: evSetTitle,
-      SET_FAVICON: evSetFavicon,
-      SET_TITLE_LAUNCHPAD: evSetTitleLaunchpad,
-      SET_FOCUS: evSetFocus,
-      SCROLL_TO: evScrollTo,
-      SCROLL_INTO_VIEW: evScrollIntoView,
-      START_TIMER: evStartTimer,
-      KEYBOARD_SET_MODE: evSetInputMode,
-      KEYBOARD_SHORTCUT: evKeyboardShortcut,
-      Z2UI5: evZ2ui5Custom,
-      WIZARD_SET_NEXT_STEP: evWizardSetNextStep,
-      PLAY_AUDIO: evPlayAudio,
-      SMART_VARIANT_INIT: evSmartVariantInit,
-      FILTER_BAR_VARIANT_INIT: evFilterBarVariantInit,
-      CONTROL_BY_ID: evControlCallById,
-      CONTROL_GLOBAL: evControlCall,
-      BINDING_CALL: evBindingCall,
-    };
 
     // Entry point called by View1.controller's eF().
     function execute(oController, args) {
@@ -1963,6 +76,58 @@ sap.ui.define(
       return handler(oController, args);
     }
 
-    return { execute, executeSystem };
+    // Run one SYSTEM action from the response's T_SYSTEM list. A system
+    // action is always framework-generated and therefore always a JSON
+    // array - there are no legacy formats here - and errors propagate: a
+    // failing view display has to reach _processAfterRendering, which turns
+    // it into the fatal overlay instead of leaving the app half-built.
+    function runSystem(item, oController) {
+      let args;
+      try {
+        args = JSON.parse(item);
+      } catch (e) {
+        Lib.logError(`systemJs: '${item}' is no action payload`, e);
+        return undefined;
+      }
+      if (!Array.isArray(args)) {
+        Lib.logError(`systemJs: '${item}' is no action payload`);
+        return undefined;
+      }
+      return executeSystem(oController, args);
+    }
+
+    // Run one APP follow-up action / custom-JS snippet from the response's
+    // T_CUSTOM list.
+    // Format A:  a JSON array ["EVENT", ...args] - the structured form the
+    //            backend (z2ui5_cl_core_srv_event=>get_event_client_json)
+    //            emits for every framework follow-up action. Pure data,
+    //            serialized and escaped entirely in ABAP; dispatched via
+    //            oController.eF( ) after a single JSON.parse - no code is
+    //            parsed or evaluated on this path.
+    // Formats B/C: legacy app-authored snippets - see actions/LegacyCustomJs.
+    function runCustom(item, oController) {
+      try {
+        const snippet = item.trim();
+        if (snippet.startsWith("[")) {
+          // JSON array -> structured follow-up action. A raw-JS expression
+          // that merely starts with "[" is no JSON array, so it fails the
+          // parse and falls through to the legacy formats.
+          try {
+            const args = JSON.parse(snippet);
+            if (Array.isArray(args)) {
+              oController.eF(...args);
+              return;
+            }
+          } catch {
+            // not JSON - keep going with the legacy formats
+          }
+        }
+        LegacyCustomJs.run(item, oController);
+      } catch (e) {
+        Lib.logError("customJs: execution failed", e);
+      }
+    }
+
+    return { execute, executeSystem, runSystem, runCustom };
   },
 );
