+++ src/01/03/z2ui5_cl_app_legacycustomjs_js.clas.xml
@@ -3,9 +3,9 @@
  <asx:abap xmlns:asx="http://www.sap.com/abapxml" version="1.0">
   <asx:values>
    <VSEOCLASS>
-    <CLSNAME>Z2UI5_CL_APP_MESSAGES_JS</CLSNAME>
+    <CLSNAME>Z2UI5_CL_APP_LEGACYCUSTOMJS_JS</CLSNAME>
     <LANGU>E</LANGU>
-    <DESCRIPT>abap2UI5 - Messages.js</DESCRIPT>
+    <DESCRIPT>abap2UI5 - LegacyCustomJs.js</DESCRIPT>
     <STATE>1</STATE>
     <CLSCCINCL>X</CLSCCINCL>
     <FIXPT>X</FIXPT>
