+++ src/01/03/z2ui5_cl_app_preload.clas.abap
@@ -52,13 +52,19 @@ CLASS z2ui5_cl_app_preload IMPLEMENTATION.
              |      "z2ui5/Component.js": function()\{{ z2ui5_cl_app_component_js=>get( ) }{ custom_js }\},| && |\n| &&
              |      "z2ui5/controller/App.controller.js": function()\{{ z2ui5_cl_app_app_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/controller/View1.controller.js": function()\{{ z2ui5_cl_app_view1_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/Browser.js": function()\{{ z2ui5_cl_app_browser_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/ControlCall.js": function()\{{ z2ui5_cl_app_controlcall_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/Launchpad.js": function()\{{ z2ui5_cl_app_launchpad_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/LegacyCustomJs.js": function()\{{ z2ui5_cl_app_legacycustomjs_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/Shortcuts.js": function()\{{ z2ui5_cl_app_shortcuts_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/Variants.js": function()\{{ z2ui5_cl_app_variants_js=>get( ) }\},| && |\n| &&
+             |      "z2ui5/core/actions/ViewOps.js": function()\{{ z2ui5_cl_app_viewops_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/AppState.js": function()\{{ z2ui5_cl_app_appstate_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/DeveloperTools.fragment.xml": '{ z2ui5_cl_app_developertool_xml=>get( ) }',| && |\n| &&
              |      "z2ui5/core/DeveloperTools.js": function()\{{ z2ui5_cl_app_developertools_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/ErrorView.js": function()\{{ z2ui5_cl_app_errorview_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/FrontendAction.js": function()\{{ z2ui5_cl_app_frontendaction_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/Lib.js": function()\{{ z2ui5_cl_app_lib_js=>get( ) }\},| && |\n| &&
-             |      "z2ui5/core/Messages.js": function()\{{ z2ui5_cl_app_messages_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/Router.js": function()\{{ z2ui5_cl_app_router_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/Server.js": function()\{{ z2ui5_cl_app_server_js=>get( ) }\},| && |\n| &&
              |      "z2ui5/core/ViewSlots.js": function()\{{ z2ui5_cl_app_viewslots_js=>get( ) }\},| && |\n| &&
